Refactor get_feature_group_url and improve DataValidationException message
