+++ bigquery.json
@@ -1,0 +1,1 @@
+{"type": "service_account", "project_id": "test"}
