+++ python/hsfs/core/external_feature_group_engine.py
@@ -13,7 +13,7 @@
 #   limitations under the License.
 #
 
-from hsfs import engine, util, client
+from hsfs import engine, util
 from hsfs import feature_group as fg
 from hsfs.client.exceptions import FeatureStoreException, DataValidationException
 from hsfs.core import feature_group_base_engine
@@ -80,11 +80,14 @@ class ExternalFeatureGroupEngine(feature_group_base_engine.FeatureGroupBaseEngin
         )
 
         if ge_report is not None and ge_report.ingestion_result == "REJECTED":
+            feature_group_url = util.get_feature_group_url(
+                feature_store_id=feature_group.feature_store_id,
+                feature_group_id=feature_group.id,
+            )
             raise DataValidationException(
                 "Data validation failed while validation ingestion policy set to strict, "
                 + f"insertion to {feature_group.name} was aborted.\n"
-                + f"Check a summary of your report at {self._get_feature_group_url(feature_group)}, "
-                + f"or download the full report from {ge_report._full_report_path}."
+                + f"You can check a summary or download your report at {feature_group_url}"
             )
 
         return (
@@ -139,14 +142,3 @@ class ExternalFeatureGroupEngine(feature_group_base_engine.FeatureGroupBaseEngin
         self._feature_group_api.update_metadata(
             feature_group, copy_feature_group, "deprecate", deprecate
         )
-
-    def _get_feature_group_url(self, feature_group):
-        sub_path = (
-            "/p/"
-            + str(client.get_instance()._project_id)
-            + "/fs/"
-            + str(feature_group.feature_store_id)
-            + "/fg/"
-            + str(feature_group.id)
-        )
-        return util.get_hostname_replaced_url(sub_path)
